requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.8.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
safetensors==0.5.3
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
